#hw15 Решение ДЗ№15

+++ hw15/src/main/java/ru/otus/java/pro/spring/app/dtos/ExecuteTransferDtoRq.java
@@ -1,4 +1,6 @@
 package ru.otus.java.pro.spring.app.dtos;
 
-public record ExecuteTransferDtoRq(String targetClientId, String sourceAccount, String targetAccount, String message, int amount) {
+import java.math.BigDecimal;
+
+public record ExecuteTransferDtoRq(String targetClientId, String sourceAccount, String targetAccount, String message, BigDecimal amount) {
 }
